requirements.txt
fastapi
jinja2
mysql-connector-python
python-dotenv
requests
tqdm
uvicorn
